Adding both description and corresponding examples within the same json file and pushing new extraction script.
