+++ Datasets/Data_Science_Meta/data_extraction.py
@@ -1,12 +1,11 @@
 import requests
 import re
 from bs4 import BeautifulSoup
-import time
 import json
-import os
+import time
 
 BASE_URL = "https://www.datasciencemeta.com/rpackages"
-OUTPUT_JSON_FILE = "r_package_examples.json"
+OUTPUT_JSON_FILE = "r_package_data.json"
 HEADERS = {
     "User-Agent": "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/100.0.4896.127 Safari/537.36"
 }
@@ -30,7 +29,6 @@ def get_package_links(base_url, top_n=100):
 
     return package_links
 
-
 def extract_github_url(package_url):
     """Extract the GitHub URL from the package's CRAN page."""
     response = requests.get(package_url, headers=HEADERS)
@@ -39,7 +37,6 @@ def extract_github_url(package_url):
         return None
 
     soup = BeautifulSoup(response.text, 'html.parser')
-
     url_row = soup.find('td', string=re.compile(r'URL:', re.I))
 
     if url_row:
@@ -51,7 +48,6 @@ def extract_github_url(package_url):
                     return url
 
     return None
-
 
 def construct_man_folder_url(github_url):
     """Construct the 'man' folder URL from the GitHub repository URL."""
@@ -67,7 +63,6 @@ def construct_man_folder_url(github_url):
             return man_url
 
     return None
-
 
 def get_rd_files(man_folder_url):
     """Retrieve .Rd file links from the 'man' folder in GitHub UI."""
@@ -87,7 +82,6 @@ def get_rd_files(man_folder_url):
 
     return rd_files
 
-
 def extract_examples_from_rd(rd_url):
     """Extracts the full \examples{} section while handling nested braces."""
     response = requests.get(rd_url, headers=HEADERS)
@@ -102,7 +96,6 @@ def extract_examples_from_rd(rd_url):
         return None
 
     content = response.text
-
     start_match = re.search(r"\\examples\s*\{", content)
     if not start_match:
         return None
@@ -120,6 +113,20 @@ def extract_examples_from_rd(rd_url):
 
     examples_body = content[start_index:end_index - 1].strip()
     return examples_body
+
+def extract_package_description(package_url):
+    """Extracts the description of the package from its CRAN page."""
+    response = requests.get(package_url, headers=HEADERS)
+    if response.status_code != 200:
+        print(f"Failed to fetch {package_url}")
+        return None
+
+    soup = BeautifulSoup(response.text, 'html.parser')
+    description_tag = soup.find('p')
+    if description_tag:
+        return description_tag.text.strip()
+    
+    return None
 
 def main():
     package_links = get_package_links(BASE_URL)
@@ -150,25 +157,27 @@ def main():
             continue
 
         package_name = github_url.split("/")[-1]
-        extracted_data[package_name] = {}
-
+        extracted_data[package_name] = {"description": None, "examples": {}}
+        
         for rd_file in rd_files:
             function_name = rd_file.split('/')[-1].replace(".Rd", "")
             example_code = extract_examples_from_rd(rd_file)
 
             if example_code:
-                extracted_data[package_name][function_name] = example_code
+                extracted_data[package_name]["examples"][function_name] = example_code
                 print(f"Extracted examples from {function_name}")
             else:
-                print(f"No \\examples section found in {function_name}.")
+                print(f"No \examples section found in {function_name}.")
 
+        if extracted_data[package_name]["examples"]:
+            extracted_data[package_name]["description"] = extract_package_description(package_url)
+        
         time.sleep(1)
 
     with open(OUTPUT_JSON_FILE, "w", encoding="utf-8") as f:
         json.dump(extracted_data, f, indent=4)
 
-    print(f"All extracted examples saved to {OUTPUT_JSON_FILE}")
-
+    print(f"All extracted data saved to {OUTPUT_JSON_FILE}")
 
 if __name__ == "__main__":
     main()
